Test Case illustrating issue 10
I encountered issue 10 myself in my build, but in a slightly different
way. When I ran my build with the -x command (eg. 'gradle -x test
cobertura') The build failed because of cobertura. Running that command
in 'DisabledTestsTest' will cause the build to fail. If the build
succeeds then issue 10 is fixed.